Adding Pool library (without work stealing implementation)
