+++ baselines/pfsp/lib/Pool.c
@@ -1,0 +1,34 @@
+#include "Pool.h"
+#include <stdlib.h>
+
+void initSinglePool(SinglePool* pool)
+{
+  pool->elements = (Node*)malloc(CAPACITY * sizeof(Node));
+  pool->capacity = CAPACITY;
+  pool->size = 0;
+}
+
+void pushBack(SinglePool* pool, Node node)
+{
+  if (pool->size >= pool->capacity) {
+    pool->capacity *= 2;
+    pool->elements = (Node*)realloc(pool->elements, pool->capacity * sizeof(Node));
+  }
+
+  pool->elements[pool->size++] = node;
+}
+
+Node popBack(SinglePool* pool, int* hasWork)
+{
+  if (pool->size > 0) {
+    *hasWork = 1;
+    return pool->elements[--pool->size];
+  }
+
+  return (Node){0};
+}
+
+void deleteSinglePool(SinglePool* pool)
+{
+  free(pool->elements);
+}
